requirements.txt
absl-py @ file:///opt/conda/conda-bld/absl-py_1639803114343/work
aiohttp @ file:///tmp/build/80754af9/aiohttp_1646806365504/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
astor==0.8.1
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
astunparse==1.6.3
async-timeout @ file:///tmp/build/80754af9/async-timeout_1637851218186/work
attention==4.1
attrs @ file:///opt/conda/conda-bld/attrs_1642510447205/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
blinker==1.4
Bottleneck @ file:///tmp/build/80754af9/bottleneck_1648028898966/work
brotlipy==0.7.0
cachetools==5.0.0
certifi @ file:///croot/certifi_1665076670883/work/certifi
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
charset-normalizer==2.0.12
click @ file:///tmp/build/80754af9/click_1646056590078/work
coverage==6.5.0
cryptography @ file:///tmp/build/80754af9/cryptography_1633520369886/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
debugpy @ file:///tmp/build/80754af9/debugpy_1637091799509/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
dill==0.3.6
entrypoints @ file:///tmp/build/80754af9/entrypoints_1649926439650/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
filelock==3.9.0
flatbuffers @ file:///home/ktietz/cip/python-flatbuffers_1634039120618/work
fonttools==4.33.3
frozenlist @ file:///tmp/build/80754af9/frozenlist_1637767113340/work
gast @ file:///Users/ktietz/demo/mc3/conda-bld/gast_1628588903283/work
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta @ file:///Users/ktietz/demo/mc3/conda-bld/google-pasta_1630577991354/work
grpcio @ file:///tmp/build/80754af9/grpcio_1637590821884/work
h5py @ file:///tmp/build/80754af9/h5py_1608649014695/work
huggingface-hub==0.11.1
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
imbalanced-learn==0.9.1
imblearn==0.0
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1648544546694/work
ipykernel @ file:///opt/conda/conda-bld/ipykernel_1662361798230/work
ipython @ file:///croot/ipython_1668088122043/work
jedi @ file:///tmp/build/80754af9/jedi_1644297102865/work
joblib @ file:///tmp/build/80754af9/joblib_1635411271373/work
jupyter_client @ file:///croot/jupyter_client_1669040270237/work
jupyter_core @ file:///croot/jupyter_core_1668084443243/work
Keras @ file:///Users/ktietz/demo/mc3/conda-bld/keras_split_1630579936814/work
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
keras-tcn==3.4.4
kiwisolver==1.4.2
libclang==14.0.1
lightgbm==3.3.2
Markdown==3.3.6
matplotlib @ file:///tmp/build/80754af9/matplotlib-suite_1647441664166/work
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mkl-service==2.4.0
multidict @ file:///opt/conda/conda-bld/multidict_1640703752579/work
munkres==1.1.4
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1649847906199/work
numexpr @ file:///tmp/build/80754af9/numexpr_1640689830538/work
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1607448156598/work
oauthlib @ file:///opt/conda/conda-bld/oauthlib_1644332107998/work
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandarallel==1.6.3
pandas==1.4.2
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.1.0
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==3.20.1
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyasn1 @ file:///Users/ktietz/demo/mc3/conda-bld/pyasn1_1629708007385/work
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
PyJWT @ file:///tmp/build/80754af9/pyjwt_1619682484438/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1635333100036/work
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2022.1
PyYAML==6.0
pyzmq @ file:///opt/conda/conda-bld/pyzmq_1657724186960/work
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.1.1
scipy==1.8.0
seaborn @ file:///tmp/build/80754af9/seaborn_1629307859561/work
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow @ file:///home/builder/ktietz/aggregate/tensorflow_recipes/ci_cpu/tensorflow-base_1614583966145/work/tensorflow_pkg/tensorflow-2.4.1-cp39-cp39-linux_x86_64.whl
tensorflow-addons==0.16.1
tensorflow-estimator @ file:///home/builder/adipietro/tf/tensorflow-estimator_1630508970172/work/tensorflow_estimator-2.6.0-py2.py3-none-any.whl
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
timm==0.6.12
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tornado @ file:///opt/conda/conda-bld/tornado_1662061693373/work
tqdm==4.64.1
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
typeguard==2.13.3
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
urllib3 @ file:///opt/conda/conda-bld/urllib3_1650639997961/work
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
Werkzeug==2.1.2
wget==3.2
wrapt @ file:///tmp/abs_c335821b-6e43-4504-9816-b1a52d3d3e1eel6uae8l/croots/recipe/wrapt_1657814400492/work
xgboost==1.5.0
yarl @ file:///tmp/build/80754af9/yarl_1606939947528/work
zipp==3.8.0
